Rename AuthenticationServer namespace to LauncherServer
